+++ posthog-react-native/src/posthog-rn.ts
@@ -44,11 +44,9 @@ export type PostHogOptions = PostHogCoreOptions & {
 
   /** Captures app lifecycle events such as Application Installed, Application Updated, Application Opened, Application Became Active and Application Backgrounded.
    * By default is false.
-   * If you're already using the 'captureLifecycleEvents' options with 'withReactNativeNavigation' or 'PostHogProvider, you don't need to set this.
-   * If this is set, this value has priority over the 'captureLifecycleEvents' option.
    * Application Installed and Application Updated events are not supported with persistence set to 'memory'.
    */
-  captureNativeAppLifecycleEvents?: boolean
+  captureAppLifecycleEvents?: boolean
 
   /**
    * Enable Recording of Session Replays for Android and iOS
@@ -158,9 +156,8 @@ export class PostHog extends PostHogCore {
         }
       }
 
-      // this value could have been inferred from the autocapture options (captureLifecycleEvents)
-      if (options?.captureNativeAppLifecycleEvents) {
-        void this.captureNativeAppLifecycleEvents()
+      if (options?.captureAppLifecycleEvents) {
+        void this.captureAppLifecycleEvents()
       }
 
       void this.persistAppVersion()
@@ -354,7 +351,6 @@ export class PostHog extends PostHogCore {
       maskAllTextInputs = true,
       maskAllImages = true,
       maskAllSandboxedViews = true,
-      maskPhotoLibraryImages = false,
       captureLog = true,
       captureNetworkTelemetry = true,
       iOSdebouncerDelayMs = 1000,
@@ -365,7 +361,6 @@ export class PostHog extends PostHogCore {
       maskAllTextInputs,
       maskAllImages,
       maskAllSandboxedViews,
-      maskPhotoLibraryImages,
       captureLog,
       captureNetworkTelemetry,
       iOSdebouncerDelayMs,
@@ -494,15 +489,13 @@ export class PostHog extends PostHogCore {
     }
   }
 
-  private async captureNativeAppLifecycleEvents(): Promise<void> {
+  private async captureAppLifecycleEvents(): Promise<void> {
     const appBuild = this._appProperties.$app_build
     const appVersion = this._appProperties.$app_version
 
     const isMemoryPersistence = this._persistence === 'memory'
 
-    // version and build are deprecated, but we keep them for compatibility
-    // use $app_version and $app_build instead
-    const properties: PostHogEventProperties = { ...maybeAdd('version', appVersion), ...maybeAdd('build', appBuild) }
+    const properties: PostHogEventProperties = {}
 
     if (!isMemoryPersistence) {
       const prevAppBuild = this.getPersistedProperty(PostHogPersistedProperty.InstalledAppBuild) as string | undefined
@@ -524,6 +517,7 @@ export class PostHog extends PostHogCore {
           // new app install
           this.capture('Application Installed', properties)
         } else if (prevAppBuild !== appBuild) {
+          // $app_version and $app_build are already added in the common event properties
           // app updated
           this.capture('Application Updated', {
             ...maybeAdd('previous_version', prevAppVersion),