+++ posthog-react-native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Next
 
+# 4.0.0 - 2025-06-10
+
+## Removed
+
+1. Remove `captureMode` in favor of `json` capture mode only
+2. Remove deprecated `personProperties` and `groupProperties` in favor of `setPersonPropertiesForFlags` and `setGroupPropertiesForFlags`
+3. Rename `captureNativeAppLifecycleEvents` option to `captureAppLifecycleEvents`
+   1. `captureAppLifecycleEvents` from `autocapture` is removed and replaced by `captureAppLifecycleEvents` from options
+4. Remove `version` and `build` from Lifecycle events in favor of `$app_version` and `$app_build`
+5. Remove maskPhotoLibraryImages from the SDK config
+
 # 3.16.1 – 2025-05-28
 
 ## Fixed